billService refactor, part 1
